move debug command to it's own file.
