+++ crack-core/src/commands/music/summon.rs
@@ -1,104 +1,12 @@
-use crate::commands::{cmd_check_music, do_join, help, sub_help as help, uptime_internal};
-use crate::messaging::message::CrackedMessage;
-use crate::poise_ext::MessageInterfaceCtxExt;
+use crate::commands::{cmd_check_music, do_join, help, sub_help as help};
 use crate::{
     connection::get_voice_channel_for_user_summon, errors::CrackedError, poise_ext::ContextExt,
     Context, Error,
 };
 use ::serenity::all::{Channel, ChannelId, Mentionable};
-use chrono::Duration;
-use songbird::tracks::PlayMode;
 use songbird::Call;
-use std::borrow::Cow;
 use std::sync::Arc;
-use std::{fmt, fmt::Display};
 use tokio::sync::Mutex;
-
-#[derive(Debug, Clone, Default)]
-pub struct BotStatus<'ctx> {
-    pub name: Cow<'ctx, String>,
-    pub play_mode: PlayMode,
-    //pub queue: Vec<TrackHandle>,
-    pub queue_len: usize,
-    pub current_channel: Option<ChannelId>,
-    pub uptime: Duration,
-}
-
-impl<'ctx> Display for BotStatus<'ctx> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "BotStatus {{\n\tplay_mode: {:?},\n\tqueue_len: {},\n\tcurrent_channel: {:?},\n\tuptime: {:?}\n\t}}",
-            self.play_mode, self.queue_len, self.current_channel, self.uptime
-        )
-    }
-}
-
-use poise::serenity_prelude as serenity;
-
-#[poise::command(
-    category = "Music",
-    slash_command,
-    prefix_command,
-    guild_only,
-    owners_only
-)]
-pub async fn debug(ctx: Context<'_>) -> Result<(), Error> {
-    let guild_id = ctx.guild_id().ok_or(CrackedError::GuildOnly)?;
-    let manager = songbird::get(ctx.serenity_context()).await.unwrap();
-    let _guild = ctx.guild().ok_or(CrackedError::NoGuildCached)?.clone();
-    let _user_id = ctx.get_user_id();
-
-    // Get the voice channel we're in if any.
-    let call = manager.get(guild_id);
-    let mut vc_status = match call {
-        Some(call) => {
-            let handler = call.lock().await;
-            let channel_id = handler.current_channel();
-            let _is_connected = handler.current_connection().is_some();
-            let queue = handler.queue();
-            let track = queue.current().clone();
-            let cur_user = ctx.cache().current_user().clone();
-            // let name = cur_user.name.clone();
-            let bot_name: Cow<'_, String> = Cow::Owned(cur_user.name.clone());
-
-            let bot_status = match track {
-                Some(track) => BotStatus {
-                    name: bot_name.into(),
-                    play_mode: track.get_info().await.unwrap_or_default().playing,
-                    current_channel: channel_id.map(|id| serenity::ChannelId::new(id.0.into())),
-                    queue_len: queue.clone().len(),
-                    uptime: Duration::zero(),
-                },
-                None => Default::default(),
-            };
-            bot_status
-        },
-        _ => Default::default(),
-    };
-    let uptime = match uptime_internal(ctx).await {
-        CrackedMessage::Uptime { seconds, .. } => Duration::seconds(seconds as i64),
-        _ => Duration::zero(),
-    };
-    vc_status.uptime = uptime;
-
-    let msg = vc_status.to_string();
-    ctx.send_reply(CrackedMessage::Other(msg), true).await?;
-
-    // let handler = call.lock().await;
-    // let channel_id = handler.current_channel();
-    // let is_connected = handler.current_connection().is_some();
-    // let queue = handler.queue().clone();
-
-    // let mut bot_status = BotStatus {
-    //     is_playing: handler.is_playing(),
-    //     is_paused: handler.is_paused(),
-    //     is_stopped: handler.is_stopped(),
-    //     current_channel: channel_id.map(|c| c.0),
-    // };
-
-    Ok(())
-}
 
 /// Summon the bot to your voice channel.
 #[poise::command(
